Updating module and fixing latest issue in the issue queue.
The major difference is that now the formatter comes with a settings form where you can enable if it should work in multiple mode or not.
Less filters on the formatter Views selection, so we can use any view we want as long as it is properly tagged with 'views_field_formatter'.

Updating also the default Views with contextual filters to help beginners.